+++ api/src/core/adapters/hal/getHalSoftware.test.ts
@@ -14,6 +14,7 @@ describe("HAL", () => {
                 "description": { "en": "-", "fr": undefined },
                 "developers": [
                     {
+                        "affiliatedStructure": [],
                         "id": "0000-0002-9777-5560",
                         "name": "Morane Gruenpeter",
                         "url": "https://orcid.org/0000-0002-9777-5560"
